[refactor][FIB-554] Lift the lamella list out of the shared widget library
custom_widgets.py is imported by 78 files -- it is the base the rest of the UI is
built on -- and it was the only part of fibsem/ui that reached into
fibsem.applications.autolamella. So the shared library dragged the application
into every one of its consumers.

The issue framed this as three incidental couplings to invert. It is not: they
are one contiguous 355-line block, and all of it is a single feature.

    _lamella_status_text      renders lamella.task_state
    _lamella_defect_icon      renders lamella.defect
    _LamellaRow               a row, and the defect menu that WRITES DefectState
    LamellaNameListWidget     the public widget

Inverting the two display helpers would have worked -- pass in a resolved colour
and label. _LamellaRow._on_defect_clicked constructs DefectState(state=...) and
assigns lamella.defect, so inverting that means passing in a menu spec and a
setter callback: a lot of machinery so an application widget can pretend to be
generic. Moved to its own module instead, which is what it is.

Also corrects the issue's other premise. The function-body imports were NOT a
circular-import workaround -- autolamella.structures does not import fibsem.ui,
both directions import cleanly, and five sibling widgets already import it at
module level. They were making the shared library degrade when the application
package is absent, which is a fair goal reached the wrong way. In a
lamella-specific module the imports are plain module-level ones and the
try/except ImportError arms are gone.

Consumers updated rather than re-exported from custom_widgets: a compatibility
re-export would keep the import edge and defeat the point.

Not moved to fibsem/applications/autolamella/ui/, where it belongs eventually.
Three of its six consumers (the minimap, the coincidence viewer, the spot burn
widget) are in fibsem/ui/, so moving it now would create three new
generic-to-application imports instead of removing any. That is FIB-559, after
FIB-407.

custom_widgets.py 1289 -> 920 lines, with no reference to autolamella. Six
imports it no longer needs are pruned with it.
