+++ fibsem/ui/widgets/tests/test_lamella_name_list_widget.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from PyQt5.QtWidgets import QApplication, QVBoxLayout, QWidget
 
 from fibsem.applications.autolamella.structures import Experiment
-from fibsem.ui.widgets.custom_widgets import LamellaNameListWidget
+from fibsem.ui.widgets.lamella_name_list_widget import LamellaNameListWidget
 
 EXPERIMENT_PATH = Path(
     "/home/patrick/github/fibsem/fibsem/applications/autolamella/log/"
